Additional vagrant machines testing implemented.
What's intended with this addition is to be able to test salt's test suite on multiple machines at once and calculate more accurately what is it's coverage since some modules are specific to some systems.

In order to achieve the remote testing, a `runtests` module was created. This module is **NOT** supposed to be used in sensible systems. It's a destructive module because in order to combine the coverage data from multiple machines, we need to re-create the same salt source path on the vagrant machines. This is the only way we can "aggregate" the coverage data correctly, or, open the coverage data file and switch paths, also not that clean.